+++ adapter_test/test_helper.exs
@@ -21,6 +21,8 @@ version_query =
 [[version]] = repo.query!(version_query).rows
 IO.puts("Adapter probe engine: #{version}")
 
-if repo.__adapter__() == Ecto.Adapters.SQLite3 do
+if repo.__adapter__() in [Ecto.Adapters.SQLite3, Ecto.Adapters.MyXQL] do
+  {:ok, _} = Ithibati.AdapterIdentityRepo.start_link()
+  Ecto.Adapters.SQL.Sandbox.mode(Ithibati.AdapterIdentityRepo, :auto)
   Code.require_file("../test/support/test_credentials.ex", __DIR__)
 end